+++ packages/nuxt/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @posthog/nuxt
 
+## 1.5.40
+
+### Patch Changes
+
+- Updated dependencies [[`4580975`](https://github.com/PostHog/posthog-js/commit/4580975b499069a60567e2844cfb0d8970ca641d)]:
+  - posthog-js@1.345.2
+
+## 1.5.39
+
+### Patch Changes
+
+- Updated dependencies [[`de43d70`](https://github.com/PostHog/posthog-js/commit/de43d70e5d94f74cf58745695968eee09fbc64b6)]:
+  - posthog-js@1.345.1
+
+## 1.5.38
+
+### Patch Changes
+
+- Updated dependencies [[`fe8090c`](https://github.com/PostHog/posthog-js/commit/fe8090c00f0122ed4aad37465f43480c50392506), [`90db5d4`](https://github.com/PostHog/posthog-js/commit/90db5d48d7ef0bff0ece31fa9a1047c10b76aa55)]:
+  - posthog-js@1.345.0
+
+## 1.5.37
+
+### Patch Changes
+
+- Updated dependencies [[`afdf746`](https://github.com/PostHog/posthog-js/commit/afdf7467ecff7bf8ed8f1645de0d370808c3ee9a), [`d578824`](https://github.com/PostHog/posthog-js/commit/d578824395ceba3b854970c2a7723e97466d9e9d)]:
+  - posthog-js@1.344.0
+  - @posthog/core@1.21.0
+  - posthog-node@5.24.14
+
+## 1.5.36
+
+### Patch Changes
+
+- Updated dependencies [[`846a7f2`](https://github.com/PostHog/posthog-js/commit/846a7f23044cb65c458ee31cdbd6e8c93a73d3bf)]:
+  - posthog-node@5.24.13
+
+## 1.5.35
+
+### Patch Changes
+
+- Updated dependencies [[`e055f9a`](https://github.com/PostHog/posthog-js/commit/e055f9a344d7c11309c56444383f79df335a5c51)]:
+  - @posthog/core@1.20.2
+  - posthog-js@1.343.2
+  - posthog-node@5.24.12
+
 ## 1.5.34
 
 ### Patch Changes