+++ packages/webpack-plugin/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @posthog/webpack-plugin
 
+## 1.2.20
+
+### Patch Changes
+
+- Updated dependencies [[`d578824`](https://github.com/PostHog/posthog-js/commit/d578824395ceba3b854970c2a7723e97466d9e9d)]:
+    - @posthog/core@1.21.0
+
+## 1.2.19
+
+### Patch Changes
+
+- Updated dependencies [[`e055f9a`](https://github.com/PostHog/posthog-js/commit/e055f9a344d7c11309c56444383f79df335a5c51)]:
+    - @posthog/core@1.20.2
+
 ## 1.2.18
 
 ### Patch Changes